✨ add watch subcommand to GitHub command for monitoring repository notifications
